Split the implementation of `ExecutionPlan` into separate implementations for mutable and finalized plans.
