+++ subprojects/core/src/test/groovy/org/gradle/execution/plan/DefaultExecutionPlanTest.groovy
@@ -40,6 +40,7 @@ import static org.gradle.util.internal.WrapUtil.toList
 
 class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
     DefaultExecutionPlan executionPlan
+    DefaultFinalizedExecutionPlan finalizedPlan
 
     def taskNodeFactory = new TaskNodeFactory(thisBuild, Stub(DocumentationRegistry), Stub(BuildTreeWorkGraphController), nodeValidator, new TestBuildOperationExecutor())
     def dependencyResolver = new TaskDependencyResolver([new TaskNodeDependencyResolver(taskNodeFactory)])
@@ -682,7 +683,7 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
         executedTasks == [a]
 
         when:
-        executionPlan.collectFailures(failures)
+        finalizedPlan.collectFailures(failures)
 
         then:
         failures == [exception]
@@ -696,14 +697,14 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
         when:
         addToGraphAndPopulate([a, b])
         coordinator.withStateLock {
-            executionPlan.cancelExecution()
+            finalizedPlan.cancelExecution()
         }
 
         then:
         executedTasks == []
 
         when:
-        executionPlan.collectFailures(failures)
+        finalizedPlan.collectFailures(failures)
 
         then:
         failures.size() == 1
@@ -715,16 +716,16 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
         RuntimeException failure = new RuntimeException()
         Task a = task("a", failure: failure)
         Task b = task("b")
-        addToGraphAndPopulate([a, b])
 
         when:
         executionPlan.setContinueOnFailure(true)
+        addToGraphAndPopulate([a, b])
 
         then:
         executedTasks == [a, b]
 
         when:
-        executionPlan.collectFailures(failures)
+        finalizedPlan.collectFailures(failures)
 
         then:
         failures == [failure]
@@ -735,16 +736,16 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
         RuntimeException failure = new RuntimeException()
         Task a = task("a", failure: failure)
         Task b = task("b", (orderingRule): [a])
-        addToGraphAndPopulate([a, b])
 
         when:
         executionPlan.setContinueOnFailure(true)
+        addToGraphAndPopulate([a, b])
 
         then:
         executedTasks == [a, b]
 
         when:
-        executionPlan.collectFailures(failures)
+        finalizedPlan.collectFailures(failures)
 
         then:
         failures == [failure]
@@ -760,16 +761,16 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
         final Task b = task("b", dependsOn: [a])
         final Task c = task("c")
         final Task d = task("d", dependsOn: [b, c])
-        addToGraphAndPopulate([d])
 
         when:
         executionPlan.setContinueOnFailure(true)
+        addToGraphAndPopulate([d])
 
         then:
         executedTasks == [a, c]
 
         when:
-        executionPlan.collectFailures(failures)
+        finalizedPlan.collectFailures(failures)
 
         then:
         failures == [failure]
@@ -795,7 +796,7 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
         executedTasks == [c]
 
         when:
-        executionPlan.collectFailures(failures)
+        finalizedPlan.collectFailures(failures)
 
         then:
         failures == []
@@ -935,7 +936,7 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
         Task b = task("b")
         addToGraphAndPopulate([a, b])
         coordinator.withStateLock {
-            executionPlan.cancelExecution()
+            finalizedPlan.cancelExecution()
         }
 
         when:
@@ -1046,7 +1047,7 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
         addToGraphAndPopulate([task3])
         executes(task1, task2, task3)
         def failures = []
-        executionPlan.collectFailures(failures)
+        finalizedPlan.collectFailures(failures)
 
         then:
         failures == [failure]
@@ -1086,7 +1087,7 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
 
     private void populateGraph() {
         executionPlan.determineExecutionPlan()
-        executionPlan.finalizePlan()
+        finalizedPlan = executionPlan.finalizePlan()
     }
 
     void executes(Task... expectedTasks) {
@@ -1122,9 +1123,9 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
     List<Node> getExecutedNodes() {
         def nodes = []
         coordinator.withStateLock {
-            while (executionPlan.executionState() != WorkSource.State.NoMoreWorkToStart) {
-                assert executionPlan.executionState() == WorkSource.State.MaybeWorkReadyToStart // There should always be a node ready to start when executing sequentially
-                def selection = executionPlan.selectNext()
+            while (finalizedPlan.executionState() != WorkSource.State.NoMoreWorkToStart) {
+                assert finalizedPlan.executionState() == WorkSource.State.MaybeWorkReadyToStart // There should always be a node ready to start when executing sequentially
+                def selection = finalizedPlan.selectNext()
                 if (selection.noMoreWorkToStart) {
                     break
                 }
@@ -1132,10 +1133,10 @@ class DefaultExecutionPlanTest extends AbstractExecutionPlanSpec {
                 def nextNode = selection.item
                 assert !nextNode.isComplete()
                 nodes << nextNode
-                executionPlan.finishedExecuting(nextNode, null)
+                finalizedPlan.finishedExecuting(nextNode, null)
             }
-            assert executionPlan.executionState() == WorkSource.State.NoMoreWorkToStart
-            assert executionPlan.allExecutionComplete()
+            assert finalizedPlan.executionState() == WorkSource.State.NoMoreWorkToStart
+            assert finalizedPlan.allExecutionComplete()
         }
         return nodes
     }
