+++ subprojects/core/src/test/groovy/org/gradle/execution/taskgraph/DefaultTaskExecutionGraphSpec.groovy
@@ -44,6 +44,7 @@ import org.gradle.execution.plan.AbstractExecutionPlanSpec
 import org.gradle.execution.plan.DefaultExecutionPlan
 import org.gradle.execution.plan.DefaultPlanExecutor
 import org.gradle.execution.plan.ExecutionNodeAccessHierarchy
+import org.gradle.execution.plan.FinalizedExecutionPlan
 import org.gradle.execution.plan.LocalTaskNode
 import org.gradle.execution.plan.Node
 import org.gradle.execution.plan.NodeExecutor
@@ -125,10 +126,10 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
         def a = brokenTask("a", failure)
 
         given:
-        populate([a])
+        def finalizedPlan = populate([a])
 
         when:
-        def result = taskGraph.execute(executionPlan)
+        def result = taskGraph.execute(finalizedPlan)
 
         then:
         result.failures == [failure]
@@ -296,9 +297,9 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
         Task c = task("c")
 
         when:
-        populate([b])
+        def finalizedPlan = populate([b])
         taskGraph.allTasks
-        taskGraph.execute(executionPlan)
+        taskGraph.execute(finalizedPlan)
 
         then:
         // tests existing behaviour, not desired behaviour
@@ -310,8 +311,8 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
         def plan2 = newExecutionPlan()
         plan2.addEntryTasks([c])
         plan2.determineExecutionPlan()
-        plan2.finalizePlan()
-        taskGraph.populate(plan2)
+        def finalizedPlan2 = plan2.finalizePlan()
+        taskGraph.populate(finalizedPlan2)
 
         then:
         !taskGraph.hasTask(":a")
@@ -335,15 +336,15 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
         def plan2 = newExecutionPlan()
         plan2.addEntryTasks([c])
         plan2.determineExecutionPlan()
-        plan2.finalizePlan()
-        taskGraph.populate(plan2)
+        def finalizedPlan2 = plan2.finalizePlan()
+        taskGraph.populate(finalizedPlan2)
 
         then:
         taskGraph.allTasks == [c]
 
         when:
         executedTasks.clear()
-        def result2 = taskGraph.execute(plan2)
+        def result2 = taskGraph.execute(finalizedPlan2)
 
         then:
         executedTasks == [c]
@@ -377,13 +378,12 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
             Stub(ServiceRegistry)
         )
         TaskExecutionGraphListener listener = Mock(TaskExecutionGraphListener)
-        Task a = task("a")
 
         when:
         taskGraph.addTaskExecutionGraphListener(listener)
-        executionPlan.addEntryTasks([a])
-        taskGraph.populate(executionPlan)
-        taskGraph.execute(executionPlan)
+        def finalizedPlan = Stub(FinalizedExecutionPlan)
+        taskGraph.populate(finalizedPlan)
+        taskGraph.execute(finalizedPlan)
 
         then:
         1 * listener.graphPopulated(_)
@@ -392,8 +392,8 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
         1 * planExecutor.process(_, _)
 
         when:
-        taskGraph.populate(executionPlan)
-        taskGraph.execute(executionPlan)
+        taskGraph.populate(finalizedPlan)
+        taskGraph.execute(finalizedPlan)
 
         then:
         0 * listener._
@@ -419,9 +419,9 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
         when:
         taskGraph.whenReady(closure)
         taskGraph.whenReady(action)
-        executionPlan.addEntryTasks([a])
-        taskGraph.populate(executionPlan)
-        taskGraph.execute(executionPlan)
+        def finalizedPlan = Stub(FinalizedExecutionPlan)
+        taskGraph.populate(finalizedPlan)
+        taskGraph.execute(finalizedPlan)
 
         then:
         1 * closure.call()
@@ -438,9 +438,9 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
         }
 
         when:
-        def plan2 = newExecutionPlan()
-        taskGraph.populate(plan2)
-        taskGraph.execute(plan2)
+        def finalizedPlan2 = Stub(FinalizedExecutionPlan)
+        taskGraph.populate(finalizedPlan2)
+        taskGraph.execute(finalizedPlan2)
 
         then:
         0 * closure._
@@ -528,13 +528,13 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
 
         when:
         executionPlan.useFilter(spec)
-        populate([a, b])
+        def finalizedPlan = populate([a, b])
 
         then:
         taskGraph.allTasks == [b]
 
         when:
-        def result = taskGraph.execute(executionPlan)
+        def result = taskGraph.execute(finalizedPlan)
 
         then:
         executedTasks == [b]
@@ -553,13 +553,13 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
 
         when:
         executionPlan.useFilter(spec)
-        populate([c])
+        def finalizedPlan = populate([c])
 
         then:
         taskGraph.allTasks == [b, c]
 
         when:
-        def result = taskGraph.execute(executionPlan)
+        def result = taskGraph.execute(finalizedPlan)
 
         then:
         executedTasks == [b, c]
@@ -586,28 +586,29 @@ class DefaultTaskExecutionGraphSpec extends AbstractExecutionPlanSpec {
         failures == [failure]
     }
 
-    private void populate(List<Task> tasks) {
+    private FinalizedExecutionPlan populate(List<Task> tasks) {
         executionPlan.addEntryTasks(tasks)
         executionPlan.determineExecutionPlan()
-        executionPlan.finalizePlan()
-        taskGraph.populate(executionPlan)
+        def finalizedPlan = executionPlan.finalizePlan()
+        taskGraph.populate(finalizedPlan)
+        return finalizedPlan
     }
 
     void populateAndExecute(List<Task> tasks) {
-        populate(tasks)
+        def finalizedPlan = populate(tasks)
         executedTasks.clear()
         failures.clear()
-        def result = taskGraph.execute(executionPlan)
+        def result = taskGraph.execute(finalizedPlan)
         failures.addAll(result.failures)
     }
 
     void execute() {
         executionPlan.determineExecutionPlan()
-        executionPlan.finalizePlan()
-        taskGraph.populate(executionPlan)
+        def finalizedPlan = executionPlan.finalizePlan()
+        taskGraph.populate(finalizedPlan)
         executedTasks.clear()
         failures.clear()
-        def result = taskGraph.execute(executionPlan)
+        def result = taskGraph.execute(finalizedPlan)
         failures.addAll(result.failures)
     }
